+++ packages/store-ui/src/deprecated/Banner/index.tsx
@@ -2,7 +2,7 @@
 import { jsx, Image } from 'theme-ui'
 import type { FC } from 'react'
 
-import LocalizedLink from '../../LocalizedLink'
+import LocalizedLink from '../LocalizedLink'
 
 type Props = {
   src: string
