+++ packages/store-ui/src/deprecated/SearchFilter/theme/index.ts
@@ -1,6 +1,6 @@
 import type { SxStyleProp } from 'theme-ui'
 
-import { createTheme } from '../../createTheme'
+import { createTheme } from '../../../createTheme'
 import { accordionTheme } from './accordion'
 
 export const searchFilterTheme: SxStyleProp = createTheme(accordionTheme)
